Add cups/printers.conf{,.0} to ignore list
Sometime between 2012 and 2024 cups placed management of the
cups/printers.conf and cups/printers.conf.0 files under the direct
management of cups-daemon / cupsd(8).

Under Debian 12 with an /etc/ under management by etckeeper, and with
an HP Envy 4500 series on the home wifi network, the result is a trail
of daily automatic commits wherein the HP printer goes to sleep,
powers back on, assigns itself a fresh UUID, and causes cups to
increment the NextPrinterId value (and of course to log the renewed
UUID as a change to the file).

This Brownian motion of the configuration file is of little value, and
since this file is now a serialization of the dynamic internal state
of the cups daemon rather than the expression of an intentional config
change by an administrator, I would judge changes in this file newly
worthy of being ignored by default.