Include cron.daily back in etckeeper and avoid running when etckeeper.timer is enabled
Include cron.daily back in etckeeper. It was removed when the debian/
directory was removed from the upsteam source. That made documentation that
says etckeeper includes a cron job wrong. It also broke the etckeeper.spec.

cron.daily: Avoid doing anything when etckeeper.timer is enabled.

Makefile: Install cron.daily on debian based systems. This will conflict
with debian/cron.daily in the debian package, but I expect they will remove
that, or otherwise deal with it.